cogs(music): Store whole queue in DB
The first track is enqueued instantly. When the track finishes playing, the next track in queue is read from the DB and fed to Lavalink just in time.

Since Lavalink has to perform a YouTube search for each track, this allows us to instantly begin playing playlists of any length by deferring the second and following track queries to when they are about to play instead of having Lavalink perform all queries at once.

This also allows us to retain queue information even when the bot goes offline, and this also gives us more control over things like shuffle and sorting methods in the future.